ecpri: added nof_corrupted_messages to the eCPRI decoder metrics

+++ include/ocudu/ofh/receiver/ofh_receiver_data_flow_metrics.h
@@ -13,11 +13,11 @@ struct rx_data_flow_perf_metrics {
   float message_unpacking_max_latency_us;
   float message_unpacking_avg_latency_us;
 
-  /// Number of dropped messages by the data flow.
+  /// Number of dropped messages detected by the data flow.
   unsigned nof_dropped_messages;
 
   /// Number of corrupted messages by the data flow.
-  unsigned nof_corrupted_messages;
+  uint64_t nof_corrupted_messages;
 
   /// CPU usage in microseconds of the message processing in the receiver.
   float cpu_usage_us;
